# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API com implementação usando Clean Architeture",
        "title": "API-MODELO-CLEAN-ARCH",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "site.com.br:8888",
    "basePath": "/api/modelo/",
    "paths": {
        "/cliente/get01/{cpf}": {
            "get": {
                "description": "get cliente by CPF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get cliente by CPF",
                "operationId": "get-cliente-by-cpf",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cliente CPF",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/router.ClienteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "router.ClienteResponse": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "dt_nasc": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "router.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}